requirements.txt
Django>=4.2,<5.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0
python-dotenv>=1.0.0
pymysql>=1.1.0
pymilvus>=2.3.0
django-cors-headers>=4.3.0

# 测试框架依赖
pytest>=7.0
pytest-html>=3.0
pytest-json>=0.4
requests>=2.28
allure-pytest>=2.13

# 文档解析
python-docx>=0.8.11
PyPDF2>=3.0.0
